LPS-146271 Move row custom styles to general panel
