+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/browser/components/page-structure/components/item-configuration-panels/RowGeneralPanel.js
@@ -14,8 +14,9 @@
 
 import ClayForm, {ClayCheckbox, ClaySelectWithOption} from '@clayui/form';
 import PropTypes from 'prop-types';
-import React from 'react';
+import React, {useMemo} from 'react';
 
+import {COLUMN_SIZE_MODULE_PER_ROW_SIZES} from '../../../../../../app/config/constants/columnSizes';
 import {VIEWPORT_SIZES} from '../../../../../../app/config/constants/viewportSizes';
 import {
 	useDispatch,
@@ -24,6 +25,8 @@ import {
 import selectSegmentsExperienceId from '../../../../../../app/selectors/selectSegmentsExperienceId';
 import updateItemConfig from '../../../../../../app/thunks/updateItemConfig';
 import updateRowColumns from '../../../../../../app/thunks/updateRowColumns';
+import {deepEqual} from '../../../../../../app/utils/checkDeepEqual';
+import {getResponsiveColumnSize} from '../../../../../../app/utils/getResponsiveColumnSize';
 import {getResponsiveConfig} from '../../../../../../app/utils/getResponsiveConfig';
 import {useId} from '../../../../../../app/utils/useId';
 import {getLayoutDataItemPropTypes} from '../../../../../../prop-types/index';
@@ -35,8 +38,35 @@ const ROW_CONFIGURATION_IDENTIFIERS = {
 	numberOfColumns: 'numberOfColumns',
 };
 
+const CUSTOM_ROW = 'custom';
+
+const MODULES_PER_ROW_OPTIONS = [
+	[1],
+	[1, 2],
+	[1, 3],
+	[1, 2, 4],
+	[1, 2, 5],
+	[1, 2, 3, 6],
+];
+const MODULES_PER_ROW_OPTIONS_WITH_CUSTOM = MODULES_PER_ROW_OPTIONS.map(
+	(option) => [CUSTOM_ROW, ...option]
+);
+
+const VERTICAL_ALIGNMENT_OPTIONS = [
+	{label: Liferay.Language.get('top'), value: 'top'},
+	{label: Liferay.Language.get('middle'), value: 'middle'},
+	{label: Liferay.Language.get('bottom'), value: 'bottom'},
+];
+
+const ROW_STYLE_IDENTIFIERS = {
+	modulesPerRow: 'modulesPerRow',
+	reverseOrder: 'reverseOrder',
+	verticalAlignment: 'verticalAlignment',
+};
+
 export function RowGeneralPanel({item}) {
 	const dispatch = useDispatch();
+	const layoutData = useSelector((state) => state.layoutData);
 	const segmentsExperienceId = useSelector(selectSegmentsExperienceId);
 	const selectedViewportSize = useSelector(
 		(state) => state.selectedViewportSize
@@ -89,29 +119,156 @@ export function RowGeneralPanel({item}) {
 		);
 	};
 
+	const onCustomStylesValueSelect = (identifier, value) => {
+		let itemStyles = {[identifier]: value};
+
+		if (
+			selectedViewportSize !== VIEWPORT_SIZES.desktop &&
+			identifier !== ROW_STYLE_IDENTIFIERS.gutters
+		) {
+			itemStyles = {[selectedViewportSize]: itemStyles};
+		}
+
+		if (identifier === ROW_STYLE_IDENTIFIERS.numberOfColumns) {
+			const currentNumberOfColumns = rowConfig.numberOfColumns;
+			const newNumberOfColumns = value;
+
+			const columnsToBeModified = Math.abs(
+				newNumberOfColumns - currentNumberOfColumns
+			);
+
+			if (columnsToBeModified === 0) {
+				return;
+			}
+
+			if (item && item.itemId) {
+				dispatch(
+					updateRowColumns({
+						itemId: item.itemId,
+						numberOfColumns: newNumberOfColumns,
+						segmentsExperienceId,
+						viewportSizeId: selectedViewportSize,
+					})
+				);
+			}
+
+			return;
+		}
+
+		dispatch(
+			updateItemConfig({
+				itemConfig: itemStyles,
+				itemId: item.itemId,
+				segmentsExperienceId,
+			})
+		);
+	};
+
+	const getModulesPerRowOptionLabel = (value) => {
+		return value > 1
+			? Liferay.Language.get('x-modules-per-row')
+			: Liferay.Language.get('x-module-per-row');
+	};
+
+	const isCustomRow = useMemo(() => {
+		const {numberOfColumns} = rowConfig;
+
+		const {modulesPerRow} = getResponsiveConfig(
+			rowConfig,
+			selectedViewportSize
+		);
+
+		const columnSizes = item.children.map((columnId) => {
+			const columnSizeConfig = getResponsiveColumnSize(
+				layoutData.items[columnId].config,
+				selectedViewportSize
+			);
+
+			return columnSizeConfig;
+		});
+
+		const columnConfiguration =
+			COLUMN_SIZE_MODULE_PER_ROW_SIZES[numberOfColumns]?.[
+				modulesPerRow
+			] ?? [];
+
+		return !deepEqual(columnConfiguration, columnSizes);
+	}, [item.children, layoutData.items, rowConfig, selectedViewportSize]);
+
+	const modulesPerRowOptions = isCustomRow
+		? MODULES_PER_ROW_OPTIONS_WITH_CUSTOM
+		: MODULES_PER_ROW_OPTIONS;
+
 	return (
 		<>
-			<Select
-				configurationKey="numberOfColumns"
-				handleChange={handleConfigurationValueChanged}
-				label={Liferay.Language.get('number-of-modules')}
-				options={NUMBER_OF_COLUMNS_OPTIONS.map((option) => ({
-					label: option,
-				}))}
-				value={rowConfig.numberOfColumns}
-			/>
-
-			{rowConfig.numberOfColumns > 1 && (
+			{selectedViewportSize === VIEWPORT_SIZES.desktop && (
 				<>
-					<ClayCheckbox
-						checked={rowConfig.gutters}
-						label={Liferay.Language.get('show-gutter')}
-						onChange={({target: {checked}}) =>
-							handleConfigurationValueChanged('gutters', checked)
-						}
+					<Select
+						configurationKey="numberOfColumns"
+						handleChange={handleConfigurationValueChanged}
+						label={Liferay.Language.get('number-of-modules')}
+						options={NUMBER_OF_COLUMNS_OPTIONS.map((option) => ({
+							label: option,
+						}))}
+						value={rowConfig.numberOfColumns}
 					/>
+
+					{rowConfig.numberOfColumns > 1 && (
+						<>
+							<ClayCheckbox
+								checked={rowConfig.gutters}
+								label={Liferay.Language.get('show-gutter')}
+								onChange={({target: {checked}}) =>
+									handleConfigurationValueChanged(
+										'gutters',
+										checked
+									)
+								}
+							/>
+						</>
+					)}
 				</>
 			)}
+
+			<Select
+				configurationKey="modulesPerRow"
+				handleChange={onCustomStylesValueSelect}
+				label={Liferay.Language.get('layout')}
+				options={modulesPerRowOptions[
+					rowConfig.numberOfColumns - 1
+				].map((option) => ({
+					disabled: option === CUSTOM_ROW,
+					label:
+						option === CUSTOM_ROW
+							? Liferay.Language.get('custom')
+							: Liferay.Util.sub(
+									getModulesPerRowOptionLabel(option),
+									option
+							  ),
+					value: option,
+				}))}
+				value={isCustomRow ? CUSTOM_ROW : rowConfig.modulesPerRow}
+			/>
+
+			{rowConfig.numberOfColumns === 2 &&
+				rowConfig.modulesPerRow === 1 &&
+				!isCustomRow && (
+					<ClayCheckbox
+						checked={rowConfig.reverseOrder}
+						label={Liferay.Language.get('inverse-order')}
+						onChange={({target: {checked}}) =>
+							onCustomStylesValueSelect('reverseOrder', checked)
+						}
+					/>
+				)}
+
+			<Select
+				configurationKey="verticalAlignment"
+				handleChange={onCustomStylesValueSelect}
+				label={Liferay.Language.get('vertical-alignment')}
+				options={VERTICAL_ALIGNMENT_OPTIONS}
+				value={rowConfig.verticalAlignment || ''}
+			/>
 		</>
 	);
 }
