+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/browser/components/page-structure/selectors/selectPanels.js
@@ -220,9 +220,7 @@ export function selectPanels(activeItemId, activeItemType, state) {
 	else if (activeItem.type === LAYOUT_DATA_ITEM_TYPES.row) {
 		panelsIds = {
 			[PANEL_IDS.rowStyles]: canUpdateItemConfiguration,
-			[PANEL_IDS.rowGeneral]:
-				canUpdateItemConfiguration &&
-				state.selectedViewportSize === VIEWPORT_SIZES.desktop,
+			[PANEL_IDS.rowGeneral]: canUpdateItemConfiguration,
 		};
 	}
